Add "Lambda" to examples that use Java 8 lambdas
