+++ kafka-streams/src/test/java/io/confluent/examples/streams/WordCountLambdaIntegrationTest.java
@@ -5,19 +5,29 @@ import org.apache.kafka.clients.producer.KafkaProducer;
 import org.apache.kafka.clients.producer.Producer;
 import org.apache.kafka.clients.producer.ProducerRecord;
 import org.apache.kafka.clients.producer.RecordMetadata;
-import org.apache.kafka.common.serialization.ByteArrayDeserializer;
-import org.apache.kafka.common.serialization.ByteArraySerializer;
+import org.apache.kafka.common.serialization.Deserializer;
+import org.apache.kafka.common.serialization.LongDeserializer;
+import org.apache.kafka.common.serialization.LongSerializer;
+import org.apache.kafka.common.serialization.Serializer;
 import org.apache.kafka.common.serialization.StringDeserializer;
 import org.apache.kafka.common.serialization.StringSerializer;
 import org.apache.kafka.streams.KafkaStreams;
+import org.apache.kafka.streams.KeyValue;
 import org.apache.kafka.streams.StreamsConfig;
 import org.apache.kafka.streams.kstream.KStream;
 import org.apache.kafka.streams.kstream.KStreamBuilder;
+import org.apache.kafka.streams.kstream.KTable;
+import org.apache.kafka.streams.kstream.UnlimitedWindows;
+import org.apache.kafka.streams.kstream.Windowed;
+import org.apache.kafka.streams.kstream.internals.WindowedSerializer;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
+import java.io.File;
 import java.io.IOException;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
@@ -26,10 +36,11 @@ import java.util.concurrent.Future;
 import java.util.stream.Collectors;
 
 import io.confluent.examples.streams.kafka.EmbeddedSingleNodeKafkaCluster;
+import kafka.utils.CoreUtils;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-public class MapFunctionIntegrationTest {
+public class WordCountLambdaIntegrationTest {
 
   private static EmbeddedSingleNodeKafkaCluster cluster = null;
   private static String inputTopic = "inputTopic";
@@ -50,29 +61,52 @@ public class MapFunctionIntegrationTest {
   }
 
   @Test
-  public void shouldUppercaseTheInput() throws Exception {
-    List<String> inputValues = Arrays.asList("hello", "world");
-    List<String> expectedValues = inputValues.stream().map(String::toUpperCase).collect(Collectors.toList());
+  public void shouldCountWords() throws Exception {
+    List<String> inputValues = Arrays.asList("hello", "world", "world", "hello world");
+    List<KeyValue<String, Long>> expectedValues = Arrays.asList(
+        new KeyValue<>("hello", 1L),
+        new KeyValue<>("world", 1L),
+        new KeyValue<>("world", 2L),
+        new KeyValue<>("hello", 2L),
+        new KeyValue<>("world", 3L)
+    );
 
     //
     // Step 1: Configure and start the Streams job.
     //
-    KStreamBuilder builder = new KStreamBuilder();
-
-    // Write the input data as-is to the output topic.
-    KStream<byte[], String> input = builder.stream(inputTopic);
-
-    KStream<byte[], String> uppercased = input.mapValues(String::toUpperCase);
-    uppercased.to(outputTopic);
+    final Serializer<String> stringSerializer = new StringSerializer();
+    final Deserializer<String> stringDeserializer = new StringDeserializer();
+    final Serializer<Long> longSerializer = new LongSerializer();
+    final Deserializer<Long> longDeserializer = new LongDeserializer();
 
     Properties streamsConfiguration = new Properties();
-    streamsConfiguration.put(StreamsConfig.JOB_ID_CONFIG, "map-function-integration-test");
+    streamsConfiguration.put(StreamsConfig.JOB_ID_CONFIG, "wordcount-integration-test");
     streamsConfiguration.put(StreamsConfig.BOOTSTRAP_SERVERS_CONFIG, cluster.bootstrapServers());
     streamsConfiguration.put(StreamsConfig.ZOOKEEPER_CONNECT_CONFIG, cluster.zookeeperConnect());
-    streamsConfiguration.put(StreamsConfig.KEY_SERIALIZER_CLASS_CONFIG, ByteArraySerializer.class);
+    streamsConfiguration.put(StreamsConfig.KEY_SERIALIZER_CLASS_CONFIG, StringSerializer.class);
     streamsConfiguration.put(StreamsConfig.VALUE_SERIALIZER_CLASS_CONFIG, StringSerializer.class);
-    streamsConfiguration.put(StreamsConfig.KEY_DESERIALIZER_CLASS_CONFIG, ByteArrayDeserializer.class);
+    streamsConfiguration.put(StreamsConfig.KEY_DESERIALIZER_CLASS_CONFIG, StringDeserializer.class);
     streamsConfiguration.put(StreamsConfig.VALUE_DESERIALIZER_CLASS_CONFIG, StringDeserializer.class);
+    streamsConfiguration.put(StreamsConfig.STATE_DIR_CONFIG, "/tmp/kafka-streams");
+
+    // Remove any state from previous test runs
+    purgeLocalStreamsState(streamsConfiguration);
+
+    KStreamBuilder builder = new KStreamBuilder();
+
+    KStream<String, String> source = builder.stream(inputTopic);
+
+    KStream<String, Long> counts = source
+        .flatMapValues(value -> Arrays.asList(value.toLowerCase().split("\\W+")))
+        .map((key, value) -> new KeyValue<>(value, value))
+        // The following countByKey().toStream().map() chain can be simplified to just
+        // countByKey() once https://github.com/apache/kafka/pull/870 is merged.
+        .countByKey(UnlimitedWindows.of("Counts").startOn(0L),
+            stringSerializer, longSerializer, stringDeserializer, longDeserializer)
+        .toStream()
+        .map((windowedKey, count) -> new KeyValue<>(windowedKey.value(), count));
+
+    counts.to(outputTopic, stringSerializer, longSerializer);
 
     KafkaStreams streams = new KafkaStreams(builder, streamsConfiguration);
     streams.start();
@@ -108,15 +142,28 @@ public class MapFunctionIntegrationTest {
     //
     Properties consumerConfig = new Properties();
     consumerConfig.put("bootstrap.servers", cluster.bootstrapServers());
-    consumerConfig.put("group.id", "map-function-integration-test-standard-consumer");
+    consumerConfig.put("group.id", "wordcount-integration-test-standard-consumer");
     consumerConfig.put("auto.offset.reset", "earliest");
     consumerConfig.put("key.deserializer", "org.apache.kafka.common.serialization.StringDeserializer");
-    consumerConfig.put("value.deserializer", "org.apache.kafka.common.serialization.StringDeserializer");
+    consumerConfig.put("value.deserializer", "org.apache.kafka.common.serialization.LongDeserializer");
 
-    KafkaConsumer<String, String> consumer = new KafkaConsumer<>(consumerConfig);
+    KafkaConsumer<String, Long> consumer = new KafkaConsumer<>(consumerConfig);
     consumer.subscribe(Collections.singletonList(outputTopic));
-    List<String> actualValues = IntegrationTestUtils.readValues(consumer, inputValues.size());
-    assertThat(actualValues).isEqualTo(expectedValues);
+    List<KeyValue<String, Long>> actualValues = IntegrationTestUtils.readKeyValues(consumer);
+
+    assertThat(actualValues).containsExactlyElementsOf(expectedValues);
+  }
+
+  private static void purgeLocalStreamsState(Properties streamsConfiguration) throws IOException {
+    String path = streamsConfiguration.getProperty(StreamsConfig.STATE_DIR_CONFIG);
+    if (path != null) {
+      File node = Paths.get(path).normalize().toFile();
+      // Only purge state when it's under /tmp.  This is a safety net to prevent accidentally
+      // deleting important local directory trees.
+      if (node.getAbsolutePath().startsWith("/tmp")) {
+        CoreUtils.rm(node);
+      }
+    }
   }
 
 }
